rename product reader modules
