+++ src/nisarqa/input_product_readers/non_insar_product.py
@@ -12,7 +12,7 @@ import numpy as np
 
 import nisarqa
 
-from .nisar_base import NisarProduct
+from .nisar_product import NisarProduct
 
 objects_to_skip = nisarqa.get_all(name=__name__)
